33.4.Adding a Custom Voter{@IsGranted with a Subject}
@IsGranted with a Subject
Voters are great. And using them to centralize this kind of logic will keep your security code solid. But, there's one small thing that now seems impossible to do. First, open ArticleAdminController. We can actually shorten this to the normal $this->denyAccessUnlessGranted('MANAGE', $article).
Try it - reload the page. Access granted! This does the exact same thing as before.