33.5.Adding a Custom Voter{@IsGranted with a Subject}
But... what about using the @IsGranted() annotation?
Hmm... now there's a problem: can we use the annotation and still, somehow, pass in the Article object? Actually, yes!
Add @IsGranted(), pass it MANAGE and then a second argument: subject="article".
That's it! When you use subject=, you're allowed to pass this the same name as any of the arguments to your controller. This only works because we used the feature that automatically queries for the Article object and passes it as an argument. These two features combine perfectly. But, if you're ever in a situation where your "subject" isn't a controller argument, no worries, just use the normal denyAccessUnlessGranted() code. But, remove it in this case.
Let's... try it! Access granted! That was too easy. Go back to the voter and comment-out the author check again - let's really make sure this is working.
Now... yes! Access denied! Go put that code back.
Oh my gosh friends, we did it! We killed this tutorial! We have a great authentication system that allows both login form authentication and API authentication! We have a rich dynamic roles system and a voter system where we can control access with any custom rules. Oh, I love security! I hope you guys are feeling empowered to create your simple, complex, crazy, whatever authentication system you need. As always, if you have questions, ask us down in the comments.
Alright people, seeya next time!